// Participium/server/src/models/enums/OfficeType.ts
export enum OfficeType {
    WATER_SUPPLY = "water_supply",
    ARCHITECTURAL_BARRIERS = "architectural_barriers",
    PUBLIC_LIGHTING = "public_lighting",
    WASTE = "waste",
    ROAD_SIGNS_AND_TRAFFIC_LIGHTS = "road_signs_and_traffic_lights",
    ROADS_AND_URBAN_FURNISHINGS = "roads_and_urban_furnishings",
    PUBLIC_GREEN_AREAS_AND_PLAYGROUNDS = "public_green_areas_and_playgrounds",
    ORGANIZATION = "organization",
    OTHER = "other"
}